+++ server/backend/cmd/app/init.go
@@ -1,6 +1,10 @@
 package app
 
-import "bigbrother_server_backend/packages/common/log"
+import (
+	"bigbrother_server_backend/packages/common/log"
+	"os"
+	"path"
+)
 
 func StartInit() {
 	if err := log.DefaultLogger.AddForwarding(log.Stdout); err != nil {
@@ -9,21 +13,17 @@ func StartInit() {
 }
 
 func EndInit() {
-	// if !config.App.ShowLogs {
-	// 	if err := log.DefaultLogger.RemoveForwarding(log.Stdout); err != nil {
-	// 		panic(err.Error())
-	// 	}
-	// }
 }
 
 func InitDefaults() {
-	// config.Init()
+	// Ain logger's Init() requires the parent directory of config.Path to exist.
+	// Create it here so Init() doesn't fail.
+	parentDir := path.Dir(log.DefaultLoggerConfig.Path)
+	if err := os.MkdirAll(parentDir, 0600); err != nil {
+		appLogger.Fatal("Failed to create log parent directory", err.Error(), nil)
+	}
 
 	appLogger.Info("Initializing default logger...", nil)
-
-	// logger.DefaultLoggerConfig.Debug = config.Debug.Enabled
-	// logger.DefaultLoggerConfig.Trace = config.App.TraceLogsEnabled
-	// logger.DefaultLoggerConfig.AppInstance = config.App.ServiceID
 
 	if err := log.DefaultLogger.Init(); err != nil {
 		appLogger.Fatal("Failed to initalize default logger", err.Error(), nil)